% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungCancerRadon.R
\docType{data}
\name{lungCancerRadon}
\alias{lungCancerRadon}
\title{Exemplary person year table}
\format{
Data frame with 5 variables.
}
\source{
Pacific Northwest National Laboratory, see 
Heidenreich, W, Jacob P, Paretzke H., et al. (1999). Radiation research, 151 2, 209-17 DOI:10.2307/3579772.
}
\usage{
data(lungCancerRadon)
}
\description{
Data set of fatal lung cancer in rats exposed to Radon at Pacific Northwest National Laboratory.
For each rat, the age at start of Radon exposure, the age at end of exposure and the age at end of follow-up/death
is provided in weeks.
The Radon dose rate is given in WL (working level).
}
\examples{
data(lungCancerRadon)
lungCancerRadon[1000,]
}
\keyword{datasets}
