% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungCancerSmoking.R
\docType{data}
\name{lungCancerSmoking}
\alias{lungCancerSmoking}
\title{Exemplary person year table}
\format{
Data frame with 6 variables.
}
\source{
Fictitious data set, only for illustration and test.
}
\usage{
data(lungCancerSmoking)
}
\description{
Fictitious data set providing the number of lung cancer cases and person years 
in some population stratified according to age, age of smoking start,
age of smoking cessation, and average number of cigarettes per day during the active smoking period.
}
\examples{
data(lungCancerSmoking)
lungCancerSmoking[1000,]
}
\keyword{datasets}
