% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msce.R
\docType{package}
\name{msce-package}
\alias{msce}
\alias{msce-package}
\title{msce: Hazard of Multi-Stage Clonal Expansion Models}
\description{
Functions to calculate hazard and survival function of Multi-Stage Clonal Expansion Models used in cancer epidemiology. For the Two-Stage Clonal Expansion Model an exact solution is implemented assuming piecewise constant parameters, see Heidenreich, Luebeck, Moolgavkar (1997) \doi{10.1111/j.1539-6924.1997.tb00878.x}. Numerical solutions are provided for its extensions, see also Little, Vineis, Li (2008) \doi{10.1016/j.jtbi.2008.05.027}.
}
\author{
\strong{Maintainer}: Cristoforo Simonetto \email{cristoforo.simonetto@ph.tum.de} (\href{https://orcid.org/0000-0003-4816-3514}{ORCID})

Other contributors:
\itemize{
  \item Jan Christian Kaiser [contributor]
}

}
\keyword{internal}
