% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msce.R
\docType{package}
\name{msce}
\alias{msce}
\title{msce: A package to compute hazard and survival function for 
Multi-Stage Clonal Expansion models}
\description{
Functions \code{\link{tsce}} and \code{\link{msce_numerical}} are provided
to compute hazard and survival function for the Two- and the more general
Multi-Stage Clonal Expansion model.
Models can be evaluated for time-dependent parameters.
However, parameters are assumed to be constant within the time intervals
specified by the user.
To allow parameter fits on many strata or cohort members, the models are
implemented with RcppParallel.
For details, see the package vignette.
}
